Move mod cli to file
